simplifications structure tests
