+++ js/src/experimental-prompt-adapters/index.ts
@@ -1,0 +1,9 @@
+import { aiSDKGenerateObjectAdapter } from "./ai-sdk";
+import { openAIChatAdapter } from "./openai";
+
+export { aiSDKGenerateObjectAdapter, openAIChatAdapter };
+
+export const adapters = {
+  openAIChat: openAIChatAdapter,
+  aiSDKGenerateObject: aiSDKGenerateObjectAdapter,
+};
